+++ CHANGELOG.md
@@ -27,6 +27,8 @@
   disabled (another fix) ([rhbz#1389941])
 - Fixed Cross-site scripting (XSS) vulnerability in web UI ([CVE-2017-2661],
   [rhbz#1434111])
+- Pcs no longer allows to create a stonith resource based on an agent whose
+  name contains a colon ([rhbz#1415080])
 
 [CVE-2017-2661]: https://access.redhat.com/security/cve/CVE-2017-2661
 [rhbz#1303969]: https://bugzilla.redhat.com/show_bug.cgi?id=1303969
@@ -35,6 +37,7 @@
 [rhbz#1362493]: https://bugzilla.redhat.com/show_bug.cgi?id=1362493
 [rhbz#1378742]: https://bugzilla.redhat.com/show_bug.cgi?id=1378742
 [rhbz#1389941]: https://bugzilla.redhat.com/show_bug.cgi?id=1389941
+[rhbz#1415080]: https://bugzilla.redhat.com/show_bug.cgi?id=1415080
 [rhbz#1421702]: https://bugzilla.redhat.com/show_bug.cgi?id=1421702
 [rhbz#1434111]: https://bugzilla.redhat.com/show_bug.cgi?id=1434111